refactor: move fetching function to separate file
