+++ client/src/App.tsx
@@ -1,43 +1,21 @@
-import { Box, Typography } from '@mui/material';
 import { useEffect, useState } from 'react';
-
-export interface EventData {
-  title: string;
-  description: string;
-  date: string;
-  organizer: string;
-}
+import EventElems from './components/eventElems/eventElems';
+import { EventData } from './types/types';
+import getAllEvents from './api/server-functions';
 
 function App() {
   const [events, setEvents] = useState<EventData[]>([]);
 
-  const eventElems = events.map((event: EventData, index) => (
-    <Box key={index}>
-      <Typography>{event.title}</Typography>
-      <Typography>{event.description}</Typography>
-      <Typography>{event.organizer}</Typography>
-      <Typography>{event.date}</Typography>
-    </Box>
-  ));
-
-  useEffect(() => {
-    fetch('/events')
-      .then(response => {
-        if (!response.ok) throw new Error('Network response was not ok');
-        return response.json();
-      })
-      .then(eventsData => {
-        console.log('DATA DATA DATA', eventsData);
-        setEvents(eventsData);
-      })
-      .catch((error) => {
-        throw new Error(`Problem with getting data from server: ${error.message}`);
-      });
+  useEffect( () => {
+    (async () => {
+      const eventsData = await getAllEvents();
+      if (eventsData) setEvents(eventsData);
+    })();
   }, []);
 
   return (
     <div>
-      { events.length === 0 ? <span>Loading...</span> : <div>{eventElems}</div> }
+      { events.length === 0 ? <span>Loading...</span> : <div><EventElems eventData={events} /></div> }
       <br />
     </div>
   );
